design: 리뷰 carousel ui 수정
